Fixed and updated Waypoints and Quests as they were not properly saving and when reloading the file it was showing that the previously completed/achieved quests/waypoints were not set properly.
Also implimented better functionality of the database grabbing items. Items are now classified as "Items" with subclasses such as "Armor, Weapon, Misc"Fixed and updated Waypoints and Quests as they were not properly saving and when reloading the file it was showing that the previously completed/achieved quests/waypoints were not set properly.

Also implimented better functionality of the database grabbing items. Items are now classified as "Items" with subclasses such as "Armor, Weapon, Misc"